requirements.txt
langchain-openai==1.0.3
python-dotenv
fpdf2
psycopg2
fastapi
uvicorn